src/main/java/br/com/fiap/fiaprestaurant/reservation/application/usecases/FindAllCompletedReservationsByCustomerIdUseCase.java
package br.com.fiap.fiaprestaurant.reservation.application.usecases;

import br.com.fiap.fiaprestaurant.reservation.application.gateways.ReservationGateway;
import br.com.fiap.fiaprestaurant.reservation.domain.entity.Reservation;
import lombok.RequiredArgsConstructor;

import java.util.List;

@RequiredArgsConstructor
public class FindAllCompletedReservationsByCustomerIdUseCase {

    private final ReservationGateway reservationGateway;

    public List<Reservation> execute(long customerId){
        return reservationGateway.findAllCompletedReservationsByCustomerId(customerId);
    }

}
